+++ src/reducers/EmployeeFormReducer.js
@@ -2,10 +2,17 @@ import {
   EMPLOYEE_UPDATE
 } from '../actions/types';
 
-const INITIAL_STATE = {};
+const INITIAL_STATE = {
+  name: ``,
+  phone: ``,
+  shift: ``,
+};
 
 export default (state = INITIAL_STATE, action) => {
   switch (action.type) {
+    case EMPLOYEE_UPDATE:
+      // action.payload === { prop: 'name', value: 'Jane' }
+      return { ...state, [action.payload.prop]: action.payload.value }
     default:
       return state;
   }
